refactor: move event_router helpers to utils/webhook.py and schemas
Reorganized event_router.py for better maintainability:

- Moved WebhookConfig and EventResponse to schemas.py
- Moved helper functions to utils/webhook.py:
  - verify_signature()
  - get_webhook_config()
  - get_event_automations()
  - create_automation_run()
- event_router.py now only contains the route handler
